requirements.txt
aiohttp
pillow
python-telegram-bot
mwt
pixivpy-async
bs4